requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.8
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.6
MarkupSafe==2.1.3

# Optional for password hashing and login management
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
